chore: migrate missing document of with-geist-ui example to typescript
