+++ examples/with-geist-ui/pages/_document.tsx
@@ -1,0 +1,38 @@
+import Document, {
+  Html,
+  Head,
+  Main,
+  NextScript,
+  DocumentContext,
+  DocumentInitialProps,
+} from 'next/document'
+import { CssBaseline } from '@geist-ui/core'
+
+export default function MyDocument() {
+  return (
+    <Html>
+      <Head />
+      <body>
+        <Main />
+        <NextScript />
+      </body>
+    </Html>
+  )
+}
+
+MyDocument.getInitialProps = async (
+  ctx: DocumentContext
+): Promise<DocumentInitialProps> => {
+  const initialProps = await Document.getInitialProps(ctx)
+  const styles = CssBaseline.flush()
+
+  return {
+    ...initialProps,
+    styles: (
+      <>
+        {initialProps.styles}
+        {styles}
+      </>
+    ),
+  }
+}
